make-test-squashfs: avoid writing output files as root.
If make-test-squashfs is ran through sudo, then give ownership
to the SUDO_USER.  This is necessary as some systems we build on
have remote-mounted /home.  There, you cannot create files as root.
so even a 'cp --preserve=ownership' will fail as that will first
write the file as root and then chmod it.

So the jump through 'tee' to avoid writing files as root.