setup - cleanups, support setting git repo and git commitish to build.
A few cleanups here, but mostly the point is to allow checking out
a specific branch.  You can pass a branch or a commitish.

SQFS_GIT defaults to the upstream
  https://github.com/AgentD/squashfs-tools-ng.git
SQFS_REF defaults to origin/fixes-1.0.0
which is what we would want to track as we're using the 1.0 api.

Example:

    SQFSNG_REF=origin/master SQFSNG_GIT=/tmp/my-sqfs.git ./setup