Add a libc fsOps implementation.
The goal of this was to replace Command() calls to 'mknod', 'chmod' and
'chown' in the fakeroot case. The goal would be to call the C library
functions that actually get wrapped by fakeroot.

That way we would not have to incur cost of subprocess for every
extract.

In current implementation, this segfaults.

I got the idea and implmeenation from
https://github.com/coreos/pkg/blob/master/dlopen/dlopen_example.go